fix: Fix wildcard application-identifier in entitlement
The provisioning profile defines a list of supported entitlements which may
include wildcard identifiers. The actual application entitlement should not
contain any wildcards. This results in bugs such as broken file import on
iOS 14.

This patch only fixes wildcards in application-identifier. It does not fix
any other keys. It does not support wildcards except '*'.

As part of the change, the feature to remove get-task-allow is also moved to
unify the entitlement modification code.